Java21 (#4450)
* Initial commit including POM changes

* Converted ehcache to new format

* Modified .travis.yml

* Jakarta import corrections, EhCache config in app YML

* Fixed file upload and corrected some of the compile errors

* fixed final complilation errors

* Some correction for JSON, EhCache wip

* Expiry order changed

* Added Mail dep and commented out some cache resolver code

* Updated AuditInterceptor

* Fiexed merge with master

* Started java21 and fixed login

* Fixed the null inject and removed dep in travis

* Fixed travis java version

* Use proper protoc

* Relaced proto extensions for optional fields

* Removed some extensions from protos and updated usage

* Created a separate ehcache file for session cache and more proto corrections

* Fixed the cache conf file

* Remove StudyType and fixed study/initiative import

* Changed the proto to keep collected and protocol DTOs

* Fixed travis build and image upload

* Removed the remaining proto extends and updated search client files

* Changed magma version

* Fixed membership addition and DAR history page

* Fixed CacheResource and changed logback version

* Fixed charts, coverage for non-opal taxos and csv download

* Added legacy upgrade and chaged joda time version

* Replaced joda time with LocalDateTime

* Removed the legacy upgrade and removed required membership dto

* Fixed templates for using old fields

* Used 'double' for JSON's proto3 'int64' problem

* Removed DTO unkonwn fields

* Fixed the logs API

* Added a EhCache event logger; changed int64 -> double + ftl fix

* Use the getFileName

* Fixed Pop, DCE cache serialization and creating new contact

* Changed stats type in DTO

---------

+++ mica-core/src/main/java/org/obiba/mica/web/model/AttachmentDtos.java
@@ -132,7 +132,7 @@ class AttachmentDtos {
     if(dto.hasType()) attachment.setType(dto.getType());
     if(dto.getDescriptionCount() > 0) attachment.setDescription(localizedStringDtos.fromDto(dto.getDescriptionList()));
     if(dto.hasLang()) attachment.setLang(new Locale(dto.getLang()));
-    attachment.setSize(dto.getSize());
+    attachment.setSize((long)dto.getSize());
     if(dto.hasMd5()) attachment.setMd5(dto.getMd5());
     attachment.setJustUploaded(dto.getJustUploaded());
